Enables building without matrix fonts
Fixes bugs caused due to the disabling of matrix fonts option provided
by commit:2869565.

Adds a conditional variable (MATRIX_FONTS) whose value becomes true or
false depending on whether enable_fonts variable has true as it's
value which in turn depends whether the user has passed the
--without-fonts flag. (enable_font is true if --without-fonts flag is
not passed along with configure). The value of MATRIX_FONTS is checked
in Makefile.am and another variable BUILD_FONTS to 0 or 1.
install-data-local in Makefile.am checks for BUILD_FONTS and installs
fonts if BUILD_FONTS is set to 1.