only restore TERM when needed, and use putenv (#36)
Only when `-f` force `TERM=linux` is used and there was a change of
`$TERM`.

Since, the forcing the TERM uses `putenv(3)`, replaced `system(3)` call
to be consistent and to get rid of compilation warning of ignoring the
return value.

Note: using `setenv(3)` is cleaner, as we only needs to keep preserved
TERM value, no need to format a `"TERM=foobar"` string.  However, as the
comment L388 stated that `putenv` is more common on non-Linux, so
keeping this way.