check headers in CMake and only include either of termios.h or termio.h
CMake builds failed on basically 93fd357 [1] on OS X since termio.h
isn't present.

[1]: https://travis-ci.org/livibetter/cmatrix/builds/337114240

cmatrix.c do not need to include both termios.h and termio.h, the latter
is a wrapper on Linux, on POSIX, the former would do.

CMake should check the header files, not defining blindly, although it's
not needed as now cmatrix.c only includes one of the headers, however,
it's a good practice, but it should also halt the build process if
neither is found.