Simplify buffer size handling in va_system: use sizeof(buf)
This partially undoes commit 66a0be3 ("`va_system()`: rename buffer
variable and make buffer size a variable"). The size of the buffer can
be determined using the sizeof keyword.

Furthermore, it is not necessary to subtract 1 from the size. From the
manual page: "The functions snprintf() and vsnprintf() write at most
size bytes (including the terminating null byte ('\0')) to str."