Don't restore the TERM variable before exiting
On UNIXoids, environment variables are process-local and destroyed when
a process exits. Thus the code at the end of main that tries to restore
TERM to its previous value can never have had the desired effect.

Even after this change, running cmatrix -f does not set TERM=linux for
the calling shell.