cmatrix.c: Adds Japanese characters support
Adds classic mode which can be invoked with "c", which uses Japanese
characters instead of the regular English characters that cmatrix
outputed earlier. Done to match up with the original matrix.

The starting, ending, and range for randum number (corresponding to the
unicode character) is modified to suite the Japanese character range.
Requires appropriate Japanese fonts to make it work.

Fixes #57
Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>