Temporary solution for #83 to allow for use of config.h with cmake
This commit provides a temporary fix for #83 that caused cmatrix to
add a background color, instead of using the default terminal
background. This issue is only present in cmake builds.

CMakeLists.txt now checks to see if a `config.h` is present. If a
`config.h` is not available, it will add the preprocessor definition
EXCLUDE_CONFIG_H as it has prior. However, if a config.h is present,
we will use that instead. This allows a user to run `autoreconf -i &&
./configure` before running `cmake` and finally, `make`.